+++ Lib/Lib/utils.cpp
@@ -21,10 +21,10 @@ void transfer(string &str, int mode)
 
 void deleteAllMark(string &s, const string &mark)
 {
-    unsigned int nSize = mark.size();
+    size_t nSize = mark.size();
     while (1)
     {
-        unsigned int pos = 0;
+        size_t pos = 0;
         pos = s.find(mark, pos);
         if (pos == string::npos)
         {
